+++ shop-chart/src/main/java/com/nalidao/shopchart/repository/ShopChartRepository.java
@@ -1,11 +1,13 @@
 package com.nalidao.shopchart.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import com.nalidao.shopchart.domain.ShopChart;
 
 @Repository
-public interface ShopChartRepository extends MongoRepository<ShopChart, Long> {
+public interface ShopChartRepository extends MongoRepository<ShopChart, BigInteger> {
 
 }
